+++ packages/go_router/lib/src/route.dart
@@ -463,7 +463,8 @@ class ShellRoute extends ShellRouteBase {
         super._() {
     for (final RouteBase route in routes) {
       if (route is GoRoute) {
-        assert(route.parentNavigatorKey == null);
+        assert(route.parentNavigatorKey == null ||
+            route.parentNavigatorKey == navigatorKey);
       }
     }
   }